Start fixing the memory map
Rather than starting RAM at 0 and adding 2G to the size, we now start it
at 2G with the correct size, with ROM at 0x1000 and some patches for
unimplemented devices.

Knowledge of the memory map has been removed from the CPU
implementation, and we now use the bbl.bin boot image.